Move `compute_single_container_offsets` to channel_positioning
